Aramco_AIS_Repo_updated_with_src/Aramco_AIS_Repo/P5_Emissions_Flaring_Reduction/requirements.txt
pandas>=1.5
numpy>=1.21
scikit-learn>=1.2
ruptures>=1.1
matplotlib>=3.7
mlflow>=2.3
pyyaml>=5.4